requirements.txt
numpy
# bpy
opencv-python
Pillow
